% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.mean.confidence.R
\name{bootstrap.mean.confidence}
\alias{bootstrap.mean.confidence}
\title{Function to find a confidence interval for the mean of a data set}
\usage{
bootstrap.mean.confidence(
  mydata,
  conf.level = 0.95,
  replicates = 10000,
  alternative = "two.sided",
  na.rm = TRUE
)
}
\arguments{
\item{mydata}{use this to pass in the data set (numeric vector) you wish to build the confidence interval for.}

\item{conf.level}{defaults to 0.95 can change this to 0.99 etc to change the confidence level required. Must be a number between 0 and 1}

\item{replicates}{defaults to 10,000 this is how many resamples to take of the data set. Can increase this to get more consistent confidence intervals.}

\item{alternative}{defaults to two.sided, can also be set to 'less' or 'greater' to get sided confidence intervals. Can also use 'l' for 'less' and 'g' for 'greater'.}

\item{na.rm}{defaults to true this is whether to remove missing values from the data set. Otherwise NA values will give an error.}
}
\value{
a vector giving the lower and upper bound of the requested confidence interval
}
\description{
This function is useful for finding the confidence interval for the mean of a data set, for which the t.test command cannot be used.
}
\examples{
data(hkhw) #load the Hong Kong Heights and Weights
bootstrap.mean.confidence(hkhw$Weight.lbs) #form the 95\% confidence interval for the median weight of Hong Kong Children
}
