% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagRegressionPlots.R
\name{diagRegressionPlots}
\alias{diagRegressionPlots}
\title{Regression Diagnostics}
\usage{
diagRegressionPlots(regression.obj, cex = 1)
}
\arguments{
\item{regression.obj}{no default, is a the output from an lm command in R}

\item{cex}{defaults to 1, controls the size of the points in the scatter plots decrease if you have a large number of points}
}
\description{
Makes a few plots which may be useful in determining whether ordinary linear regression is working for a data set.
}
\examples{
data("alligator")
lm.alligator=lm(lnLength~lnWeight, data=alligator)
summary(lm.alligator)
diagRegressionPlots(lm.alligator)
}
