% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropLowFactors.R
\name{dropLowFactors}
\alias{dropLowFactors}
\title{Function to filter out levels of factors with low counts from a data frame}
\usage{
dropLowFactors(mydf, factor.column = 1, threshold = 30)
}
\arguments{
\item{mydf}{no default this is the original data frame}

\item{factor.column}{defaults to 1, tells the function which column to consider, this should be a factor}

\item{threshold}{defaults to 30, this function will drop all rows with level of the factor.column that appears less than threshold number of times}
}
\value{
new data frame with entries removed
}
\description{
This will drop any row which has an entry in the factor.column number that appears less than threshold number of times in the data.
}
\examples{
None
}
