% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterNumericOnly.R
\name{filterNumericOnly}
\alias{filterNumericOnly}
\title{Function to drop all columns of a data frame which aren't Numeric}
\usage{
filterNumericOnly(mydf)
}
\arguments{
\item{mydf}{this is the data frame you want to filter, this won't be changed at all by the function.}
}
\value{
a data frame with the non numeric columns removed
}
\description{
This function is can be used to get a reduced data frame with only numeric columns. Useful when you want to perform clustering on a data frame with
the kmeans function which only takes numerical values.
}
\examples{
data(Distel_Turtle_Data)
num.turtle<-<-filterNumericOnly(Distel_Turtle_Data)
}
