% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatBDB.R
\name{formatBDB}
\alias{formatBDB}
\title{Format Basketball Reference.com spreadsheets}
\usage{
formatBDB(mydf)
}
\arguments{
\item{mydf}{this is the data frame to format}
}
\value{
a data frame formatted
}
\description{
This will format the Basketball References CSV downloads for statistical analysis and return a new data frame.
}
\examples{
Kawhi_Leonard_2016=formatBDB(Leonard_2016_2017)
}
