% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateCorrelatedErrors.R
\name{generateCorrelatedErrors}
\alias{generateCorrelatedErrors}
\title{Generated Non-Independent (correlated) Errors}
\usage{
generateCorrelatedErrors(n = 100, sigma = 1, lag = 10)
}
\arguments{
\item{n}{defaults to 100 gives the length of the noise vector requested}

\item{sigma}{defaults to 1, gives the standard deviation of the noise requested}

\item{lag}{fdefaults to 10 gives the lag time in the correlation}
}
\value{
a vector of length n of correlated random samples
}
\description{
Makes a few lots which may be useful in determining whether linear regression is working for a data set.
}
\examples{
generateCorrelatedErrors(1000, sigma=2.0)
}
