\name{Hogwarts_Heights}
\alias{Hogwarts_Heights}
\docType{data}
\title{
Hogwarts Heights Data Set
}
\description{
A fake data set of the heights of magical creatures and students at Hogwarts Magical School.
}
\usage{data("Hogwarts_Heights")}
\format{
  A data frame with 1000 observations on the following variable.
  \describe{
    \item{\code{Heights}}{a numeric vector giving the heights of the creatures in feet. }
  }
}
\details{
NA
}
\source{
Fake data
}
\references{
%%  ~~ possibly secondary sources and usages ~~
}
\examples{
data(Hogwarts_Heights)
hist(Hogwarts_Heights)
}
\keyword{datasets}
