% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invlogit.R
\name{invlogit}
\alias{invlogit}
\title{Inverse Logit Function for Logistic Regression}
\usage{
invlogit(x)
}
\arguments{
\item{x}{input value (convert this to a probability)}
}
\value{
invlogit(x)
}
\description{
This function is used in logistic regression
}
\examples{
invlogit(10.0)
}
