\name{letter.grade.convert}
\alias{letter.grade.convert}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Convert numeric grades to letter grades A,B,C,D,F
}
\description{
Converts a vector of numeric grades to letter grades
}
\usage{
letter.grade.convert(num_grades)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{num_grades}{
vector of numeric grades
}
}
\details{
Uses 89.5 is an A (rounds up bordeline cases) for each letter grade category
}
\value{
%%  ~Describe the value returned
%%  If it is a LIST, use
%%  \item{comp1 }{Description of 'comp1'}
%%  \item{comp2 }{Description of 'comp2'}
%% ...
}
\references{
%% ~put references to the literature/web site here ~
}
\author{
Kevin M Hannay
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~


