% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit.R
\name{logit}
\alias{logit}
\title{Logit Function for Logistic Regression}
\usage{
logit(x)
}
\arguments{
\item{x}{input value}
}
\value{
logit(x)
}
\description{
This function is used in logistic regression
}
\examples{
logit(10.0)
}
