% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSimpleResidualPlot.R
\name{makeSimpleResidualPlot}
\alias{makeSimpleResidualPlot}
\title{Function to plot the residuals for a given simple OLS Problem}
\usage{
makeSimpleResidualPlot(
  alpha = 0,
  beta = 1,
  x,
  y,
  main = "",
  xlab = "x",
  ylab = "y"
)
}
\arguments{
\item{alpha}{defaults to 0, y intercept of the linear model}

\item{beta}{defaults to 1, slope of the linear model}

\item{x}{no default, this is a vector of the x values for the data points}

\item{main, }{no default, the title of the graph}

\item{xlab, }{default to y , the y axis label of the graph}
}
\value{
NULL (nothings returned)
}
\description{
This function makes a simple plot of the residuals as arrows versus a statistical model of the form y=alpha+beta*x. Will print the least
squares cost of the model.
}
\examples{
None
}
