% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normal.Interval.Calc.R
\name{Normal.Interval.Calc}
\alias{Normal.Interval.Calc}
\title{Normal Interval Calculator (a,b)}
\usage{
Normal.Interval.Calc(a, b, mean = 0, sd = 1)
}
\arguments{
\item{a}{no default gives the left side of the interval}

\item{b}{no default, gives the right side of the interval}

\item{mean, }{defaults to zero, gives the mean of the normal distribution}

\item{sd, }{defaults to one, gives the standard deviation of the normal distribution}
}
\value{
the probability of a normal random variate with the given mean and sd lying in the given interval
}
\description{
Function to give the probability of a random sample from a Normal distribution lying in a given interval (a,b)
}
\examples{
Normal.Interval.Calc(-1,1) #empirical rule approx 68\%
}
