% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perm.test.mean.R
\name{perm.test.mean}
\alias{perm.test.mean}
\title{Function to do a permutation test for a differences in means}
\usage{
perm.test.mean(x, y, resamples = 10000)
}
\arguments{
\item{x}{this is the first sample in the data set}

\item{y}{this is the second sample in the data set}

\item{resamples}{defaults to 10000 this is how many resamples to take of the data set. Can increase this to get a more exact p value}
}
\value{
the estimated p value
}
\description{
This does a nonparametric permutation test for a difference in the means between the two samples
}
\examples{
data(hkhw) #load the Hong Kong Heights and Weights
bootstrap.median.confidence(hkhw$Weight.lbs) #form the 95\% confidence interval for the median weight of Hong Kong Children
}
