% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportion.se.R
\name{proportion.se}
\alias{proportion.se}
\title{Compute the Standard Error for a Population Proportion}
\usage{
proportion.se(success, total)
}
\arguments{
\item{success}{is the number of successes in the sample}

\item{total}{is the TOTAL number of trials in the data set}
}
\value{
the estimated standard error for estimator
}
\description{
This is a short-hand function to find the standard error for a population proportion calculation
}
\examples{
proportion.se(10,25)
}
