% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbimodalNormal.R
\name{rbimodalNormal}
\alias{rbimodalNormal}
\title{Generated Random Numbers From Bimodal Normal Distribution}
\usage{
rbimodalNormal(n, mu1 = 1, mu2 = -1, sigma1 = 0.5, sigma2 = 0.5,
  p1 = 0.5)
}
\arguments{
\item{n}{no default is the number of random variates requested}

\item{mu1}{default 1.0 is the mean of the first normal dist}

\item{mu2}{default -1.0 is the mean of the second normal dist}

\item{sigma1}{default 0.5 is the standard deviation of the first normal dist}

\item{sigma2}{default 0.5 is the standard deviation of the second normal dist}

\item{p1}{default 0.5 is the weight of the first normal distibution, p2 is 1.0-p1}
}
\value{
a vector of length n of random samples from the bimodal distribution
}
\description{
Generates random numbers from two normal distributions
}
\examples{
rbimodalNormal(100)
}
