% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runExample.R
\name{runExample}
\alias{runExample}
\title{Function used to run interactive apps included in the HannayIntroStats Package}
\usage{
runExample(example)
}
\arguments{
\item{example}{the name of the app you want to run in string format}
}
\value{
None
}
\description{
This function is useful to visualize clusters of states. It uses the kmeans function to form clusters of the states and then provides a visualization
of the data. State names should be saved as the row.names of the passed in data frame. Depends on the map function from the maps package.
}
\examples{
 runExample('AssessingNormality')
}
