% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runHannayApp.R
\name{runHannayApp}
\alias{runHannayApp}
\title{Run an App}
\usage{
runHannayApp(example)
}
\arguments{
\item{example}{the name of the app you want to run in string format}
}
\value{
None
}
\description{
This can be used to run the included apps in the HannayIntroStats package.
}
\examples{
 runExample('AssessingNormality')
}
