\name{showHannayDataSets}
\alias{showHannayDataSets}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
List of Data Sets
}
\description{
Opens a window which shows a list of all the data sets in the package. Useful to find
a data set for the class project.
}
\usage{
showHannayDataSets()
}

