% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatePlot.R
\name{StatePlot}
\alias{StatePlot}
\title{Function to plot kmeans clusters in the US States}
\usage{
StatePlot(numClusters = 2, mydf, drops = NULL, scale = TRUE)
}
\arguments{
\item{numClusters}{defaults to 2, tells the function how many clusters you want to form on your data set}

\item{mydf}{this is the data frame to perform the kmeans clustering on. By default all columns of the data will be used.
this data frame must have the state names as the row names. It cannot currently handle abbreviations of the states names.}

\item{drops}{defaults to NULL (none) use this if you want to drop one of more columns from the clustering data set}

\item{scale}{defaults to TRUE. This will scale each column to have a mean of zero and standard deviation of one. This prevents the clustering algrorithm from
overweighting one column over another and can allow for easier interpretations of the results.}
}
\value{
The kmeans cluster object. It will also print the centers of the clusters to the command line.
}
\description{
This function is useful to visualize clusters of states. It uses the kmeans function to form clusters of the states and then provides a visualization
of the data. State names should be saved as the row.names of the passed in data frame. Depends on the map function from the maps package.
}
\examples{
data(crime_cluster) #load the data crime_cluster
StatePlot(2, crime_cluster) #make the plot with two clusters
}
