% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t.test.hand.R
\name{t.test.hand}
\alias{t.test.hand}
\title{T test calculator using summary statistics}
\usage{
\method{t}{test.hand}(m1, s1, n1, m2 = NULL, s2 = NULL, n2 = NULL,
  mu = 0, conf.level = 0.95, var.equal = FALSE,
  alternative = c("two.sided", "less", "greater"))
}
\arguments{
\item{m1}{the mean of data set one (numeric value)}

\item{s1}{the standard deviation of data set one (numeric value)}

\item{n1}{the number of data points in data set one, the sample size (numeric value)}

\item{m2}{the mean of data set two (numeric value)}

\item{s2}{the standard deviation of data set two (numeric value)}

\item{n2}{the number of data points in data set two, the sample size (numeric value)}

\item{mu}{a number indicating the true value of the mean (or difference in means if you are performing a two sample test).}

\item{conf.level}{confidence level of the interval.}

\item{var.equal}{a logical variable indicating whether to treat the two variances as being equal. If TRUE then the pooled variance is used to estimate the variance otherwise the Welch (or Satterthwaite) approximation to the degrees of freedom is used.}

\item{alternative}{a character string specifying the alternative hypothesis, must be one of "two.sided" (default), "greater" or "less". You can specify just the initial letter.}
}
\value{
A list with class "htest" containing the following components:
\item{statistic}{the value of the t-statistic.}
\item{parameter}{the degrees of freedom for the t-statistic.}
\item{p.value}{the p-value for the test.}
\item{conf.int}{a confidence interval for the mean appropriate to the specified alternative hypothesis.}
\item{estimate}{the estimated mean or difference in means depending on whether it was a one-sample test or a two-sample test.}
\item{null.value}{the specified hypothesized value of the mean or mean difference depending on whether it was a one-sample test or a two-sample test.}
\item{alternative}{a character string describing the alternative hypothesis.}
\item{method}{a character string indicating what type of t-test was performed.}
\item{data.name}{a character string giving the name(s) of the data.}
}
\description{
Function to form confidence intervals and perform t tests for data where we only have summary statistics. This function will perform inference for data sets where we only have the mean, sd and counts.
}
\examples{
t.test.hand(m1=10.0, s1=2.0, n1=100) #forn a 95\% confidence interval
t.test.hand(m1=10.0, s1=2.0, n1=30, m2=9.0, s2=1.0, n2=40) #perform a two sided two sample t test
}
