% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagRegressionPlots.R
\name{diagRegressionPlots}
\alias{diagRegressionPlots}
\title{Regression Diagnostics}
\usage{
diagRegressionPlots(regression.obj, x, y)
}
\arguments{
\item{regression.obj}{no default, is a the output from an lm command in R}

\item{x, }{no default gives the x values used in the regression used to make a data plot}

\item{y, }{no default gives the y values used in the regression used to make a data plot}
}
\description{
Makes a few lots which may be useful in determining whether linear regression is working for a data set.
}
\examples{
data("AmesHousing_Regression")
lm.res<-lm(SalePrice.log10~Square.Feet.log10, data=AmesHousing_Regression)
summary(lm.res)
diagRegressionPlots(lm.res)
}
